+++ server/src/main/java/com/plantity/server/repository/PlantDetailRepository.java
@@ -3,5 +3,5 @@ package com.plantity.server.repository;
 import com.plantity.server.domain.plant.detail.PlantDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlantDetailRepository extends JpaRepository<PlantDetail, Long> {
+public interface PlantDetailRepository extends JpaRepository<PlantDetail, Integer> {
 }
